Convert api module into package
This is an initial update to build a more granular
module structure for API endpoints.

* Convert `api` module into a package.
* Rename `ansible_events_ui.manager` -> `ansible_events_ui.ruleset`.
* Extract `TaskManager`, `ConnectionManager` and `UpdateManager` classes
  into a separate module `ansible_events_ui.managers`.
* Update tests structure.
* Create a sample empty module for the Activation API endpoints.
