+++ ansible_events_ui/api/activation.py
@@ -1,0 +1,8 @@
+"""Activation API endpoints."""
+from fastapi import APIRouter
+
+__all__ = ("router",)
+
+router = APIRouter()
+
+# Implement activation API here.
