+++ tests/integration/api/test_activation.py
@@ -1,0 +1,1 @@
+# Implement tests for activation API here.
