+++ tests/integration/api/test_inventory.py
